return types, 64-bit changes: ioctl cmd -> u_long, casts, etc.
SCCS-vsn: sys/kern/subr_log.c 8.2
SCCS-vsn: sys/kern/tty_tb.c 8.2
SCCS-vsn: sys/kern/tty.c 8.13
SCCS-vsn: sys/kern/kern_sig.c 8.12
SCCS-vsn: sys/kern/tty_pty.c 8.3
SCCS-vsn: sys/kern/kern_resource.c 8.7
SCCS-vsn: sys/kern/sys_socket.c 8.2
SCCS-vsn: sys/kern/vfs_bio.c 8.11
SCCS-vsn: sys/kern/uipc_usrreq.c 8.6
SCCS-vsn: sys/kern/kern_prot.c 8.8
SCCS-vsn: sys/kern/uipc_mbuf.c 8.3
SCCS-vsn: sys/kern/kern_proc.c 8.6
SCCS-vsn: sys/kern/tty_tty.c 8.3
SCCS-vsn: sys/kern/tty_compat.c 8.2
SCCS-vsn: sys/kern/vfs_subr.c 8.14
SCCS-vsn: sys/kern/tty_conf.c 8.5
SCCS-vsn: sys/kern/vfs_vnops.c 8.6

+++ usr/src/sys/kern/kern_sig.c
@@ -3,7 +3,7 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)kern_sig.c	8.11 (Berkeley) %G%
+ *	@(#)kern_sig.c	8.12 (Berkeley) %G%
  */
 
 #define	SIGPROP		/* include signal properties table */
@@ -592,7 +592,7 @@ void
 trapsignal(p, signum, code)
 	struct proc *p;
 	register int signum;
-	u_int code;
+	u_long code;
 {
 	register struct sigacts *ps = p->p_sigacts;
 	int mask;
