+++ usr/src/sys/kern/tty_tty.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)tty_tty.c	8.2 (Berkeley) %G%
+ *	@(#)tty_tty.c	8.3 (Berkeley) %G%
  */
 
 /*
@@ -88,7 +88,7 @@ cttywrite(dev, uio, flag)
 /*ARGSUSED*/
 cttyioctl(dev, cmd, addr, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t addr;
 	int flag;
 	struct proc *p;
